requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
python-ldap==2.4.15
requests==2.4.1
wsgiref==0.1.2
